Factor-out some common functionality in Batch Python scripts
