+++ .isort.cfg
@@ -2,4 +2,4 @@
 line_length = 88
 multi_line_output = 3
 include_trailing_comma = True
-known_third_party = _pytest,aenum,affine,aiohttp,alembic,async_lru,asyncpg,boto3,botocore,click,docker,errors,fastapi,fiona,geoalchemy2,geojson,gfw_pixetl,gino,gino_starlette,google,httpx,httpx_auth,logger,moto,numpy,orjson,osgeo,pandas,pendulum,pglast,psycopg2,pydantic,pyproj,pytest,rasterio,retrying,shapely,sqlalchemy,sqlalchemy_utils,starlette,tileputty,typer
+known_third_party = _pytest,aenum,affine,aiohttp,alembic,async_lru,asyncpg,aws_utils,boto3,botocore,click,docker,errors,fastapi,fiona,geoalchemy2,geojson,gfw_pixetl,gino,gino_starlette,google,httpx,httpx_auth,logger,moto,numpy,orjson,osgeo,pandas,pendulum,pglast,psycopg2,pydantic,pyproj,pytest,rasterio,retrying,shapely,sqlalchemy,sqlalchemy_utils,starlette,tileputty,typer
